pyproject.toml
[tool.poetry]
name = "browser-agent"
version = "0.1.0"
description = ""
authors = ["Pxlart Design <michael.rhema@gmail.com>"]
readme = "README.md"
package-mode = false
packages = [
    { include = "browser_agent" },
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
redis = "^5.2.1"
celery = "^5.4.0"
browser-use = "0.1.36"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
